+++ orix/tests/test_phase_list.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2018-2023 the orix developers
+# Copyright 2018-2024 the orix developers
 #
 # This file is part of orix.
 #
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-from diffpy.structure import Lattice, Structure
+from diffpy.structure import Atom, Lattice, Structure, loadStructure
 from diffpy.structure.spacegroups import GetSpaceGroup
 import numpy as np
 import pytest
@@ -333,6 +333,57 @@ class TestPhase:
         assert np.allclose(br.data, [0, 0.679, 0], atol=1e-3)
         assert np.allclose(cr.data, [0, 0, 0.714], atol=1e-3)
 
+    @pytest.mark.parametrize(
+        ["lat", "atoms"],
+        [
+            [
+                Lattice(1, 1, 1, 90, 90, 90),
+                [
+                    Atom("C", [0, 0, 0]),
+                    Atom("C", [0.5, 0.5, 0.5]),
+                    Atom("C", [0.5, 0, 0]),
+                ],
+            ],
+            [
+                Lattice(1, 1, 1, 90, 90, 120),
+                [
+                    Atom("C", [0, 0, 0]),
+                    Atom("C", [0.5, 0, 0]),
+                    Atom("C", [0.5, 0.5, 0.5]),
+                ],
+            ],
+            [
+                Lattice(1, 2, 3, 90, 90, 60),
+                [
+                    Atom("C", [0, 0, 0]),
+                    Atom("C", [0.1, 0.1, 0.6]),
+                    Atom("C", [0.5, 0, 0]),
+                ],
+            ],
+        ],
+    )
+    def test_atom_positions(self, lat, atoms):
+        structure = Structure(atoms, lat)
+        phase = Phase(structure=structure)
+        # xyz_cartn is independent of basis
+        assert np.allclose(phase.structure.xyz_cartn, structure.xyz_cartn)
+
+        # however, Phase should (in many cases) change the basis.
+        if np.allclose(structure.lattice.base, phase.structure.lattice.base):
+            # In this branch we are in the same basis & all atoms should be the same
+            for atom_from_structure, atom_from_phase in zip(structure, phase.structure):
+                assert np.allclose(atom_from_structure.xyz, atom_from_phase.xyz)
+        else:
+            # Here we have differing basis, so xyz must disagree for at least some atoms
+            disagreement_found = False
+
+            for atom_from_structure, atom_from_phase in zip(structure, phase.structure):
+                if not np.allclose(atom_from_structure.xyz, atom_from_phase.xyz):
+                    disagreement_found = True
+                    break
+
+            assert disagreement_found
+
     def test_from_cif(self, cif_file):
         """CIF files parsed correctly with space group and all."""
         phase = Phase.from_cif(cif_file)
@@ -344,6 +395,13 @@ class TestPhase:
         assert np.allclose(
             lattice.base, [[15.5, 0, 0], [0, 4.05, 0], [-1.779, 0, 6.501]], atol=1e-3
         )
+
+    def test_from_cif_same_structure(self, cif_file):
+        phase1 = Phase.from_cif(cif_file)
+        structure = loadStructure(cif_file)
+        phase2 = Phase(structure=structure)
+        assert np.allclose(phase1.structure.lattice.base, phase2.structure.lattice.base)
+        assert np.allclose(phase1.structure.xyz, phase2.structure.xyz)
 
 
 class TestPhaseList: